Avoid magic loading of elk worker?
It seems that if you attempt to load the elk worker from its standard
location of

    ../../lib/elk-worker.js

then, *even if that file doesn't exist*, and you have said "Disable cache" in
the Network tab, and even if you have restarted the browser, elkjs still
somehow magically loads some "standard" copy of that file. I know not whence it
comes.

The Network tab in the browser tools shows "Finished" for the status (instead
of 200), and shows no response headers. I don't know what's going on.

However, if you use a different name for the file, like `elk-workerFOO.js`,
then everything works normally. In particular, this allows us to rebuild
elkjs, and reload the page, and see our latest changes take effect.